organised code layout,added new dir for all table related files, stripped some logic from table file into modules.
